Move core configs to core/ folder. Leave the configs for legacy models.
PiperOrigin-RevId: 336795303
