+++ official/nlp/tasks/masked_lm.py
@@ -19,9 +19,9 @@ import dataclasses
 import tensorflow as tf
 
 from official.core import base_task
+from official.core import config_definitions as cfg
 from official.core import task_factory
 from official.modeling import tf_utils
-from official.modeling.hyperparams import config_definitions as cfg
 from official.nlp.configs import bert
 from official.nlp.configs import encoders
 from official.nlp.data import data_loader_factory
