+++ official/vision/beta/configs/image_classification.py
@@ -17,10 +17,10 @@
 import os
 from typing import List
 import dataclasses
+from official.core import config_definitions as cfg
 from official.core import exp_factory
 from official.modeling import hyperparams
 from official.modeling import optimization
-from official.modeling.hyperparams import config_definitions as cfg
 from official.vision.beta.configs import backbones
 from official.vision.beta.configs import common
