+++ official/vision/beta/configs/video_classification.py
@@ -16,10 +16,10 @@
 """Video classification configuration definition."""
 from typing import Optional, Tuple
 import dataclasses
+from official.core import config_definitions as cfg
 from official.core import exp_factory
 from official.modeling import hyperparams
 from official.modeling import optimization
-from official.modeling.hyperparams import config_definitions as cfg
 from official.vision.beta.configs import backbones_3d
 from official.vision.beta.configs import common
