+++ official/vision/beta/configs/video_classification_test.py
@@ -19,8 +19,8 @@
 from absl.testing import parameterized
 import tensorflow as tf
 
+from official.core import config_definitions as cfg
 from official.core import exp_factory
-from official.modeling.hyperparams import config_definitions as cfg
 from official.vision import beta
 from official.vision.beta.configs import video_classification as exp_cfg
